requirements.txt
autogen
python-dotenv
gitingest